Move flywheel to turret
